Remove redundant is_outlier function.
