+++ MORESCA/pipeline/__init__.py
@@ -2,7 +2,6 @@ from .batch_effect_correction import batch_effect_correction
 from .clustering import clustering
 from .diff_gene_exp import diff_gene_exp
 from .feature_selection import feature_selection
-from .is_outlier import is_outlier
 from .load_data import load_data
 from .neighborhood_graph import neighborhood_graph
 from .normalization import normalization
@@ -17,7 +16,6 @@ __all__ = [
     "clustering",
     "diff_gene_exp",
     "feature_selection",
-    "is_outlier",
     "load_data",
     "neighborhood_graph",
     "normalization",
