+++ MORESCA/tests/test_main.py
@@ -1,8 +1,9 @@
 import os
 import shutil
 import subprocess
-from pathlib import Path
 import sys
+from pathlib import Path
+
 import pytest
 import scanpy as sc
